CI,windows: create test_suites for Windows tests
Add recursive test_suite rules for all tests that
ci.bazel.io runs for Windows, and set the
top-level test_suite as the CI test target.

Doing so shortens the command line and works
around https://github.com/bazelbuild/bazel/issues/3742

I verified that the old set of tests are the same
as the new set.

Change-Id: Id8d5da3f0c03c9b8969a9f8e1e9a3096888365aa
PiperOrigin-RevId: 169242858