Always pass resource filters to aapt
It turns out that the code to filter in analysis, based on android_ide_common,
isn't as strict as aapt is.

In particular, given a language filter (like 'en'), aapt rejects language and
region resources (like 'en-rUS') but android_ide_common does not.

We could try to just patch this behavior, but it's probably indicative of
larger inconsistancies between android_ide_common and aapt. As a result, always
pass resource filters to aapt. In most cases, if we already filtered in
analysis, few if any resources will be filtered out by aapt, so we'll still
keep most of the performance gains we expected from filtering in analysis.

RELNOTES: none
PiperOrigin-RevId: 169254335