Cleanup of expansion option naming
Options that expand to other options are expansion options and the options they expand to have values that were expansions. This can be a bit confusing. Removes the isExpansion() call that is somewhat ambiguous, and forces option users to explicitly check the option definition for this information.

Also provide a parallel boolean function for implicit requirements, so that we stop querying for the length of the implicit requirement all over the place.

RELNOTES: None.
PiperOrigin-RevId: 169461566