docker_pull: rely on the main repository docker_build
The legacy docker_build was removed from bazel_tools
but we need to use it for our integration test so refer
to the one in the main repository.

Also restore the visibility, just excluding it from
bazel_tools is enough and docker_pull needs to call it
from other repository.

Change-Id: I8a8a5781859960030ae996db95fbaf9c54ec596a
PiperOrigin-RevId: 169664916