Add a global failure when a test is interrupted/cancelled.
When a timeout occurred, the current test case is interrupted and the others are cancelled. This was not reflected in any way and all tests were reported as success, even if there was a timeout and the tests were cancelled/interrupted.

Also add a status xml attribute to mark if the test was completed, cancelled or interrupted.

Fixes #3763

RELNOTES: None.
PiperOrigin-RevId: 169665622