Add an option to control building of transitive py_binary runfiles trees. , )
Currently, "bazel build :script" constructs the
bazel-bin/python-prog.runfiles tree.

We'd like to do anyway with this behavior but doing so is a
backwards-incompatible change. To facilitate a transition, this CL
adds an option --experimental_build_transitive_python_runfiles to
control the behavior. In the example above, "bazel build
--noexperimental_build_transitive_python_runfiles :script" won't
construct bazel-bin/python-prog.runfiles.

Change-Id: If4d5a84c956a0bbac9067dcf38a00e5732c450f2
PiperOrigin-RevId: 170038245