Make gzipped pkg_tar outputs repeatable
The current implementation allows the gzip module to use the current
time as a timestamp. Python2 unfortunately doesn't expose this in the
tarfile module so it's easy to overlook.

This patch separates the gzip process by creating a separate file
object manually and bypassing the gzip compression in the tarfile
module.

Fixes #1844

Change-Id: I19b265bf55314d38bdd53ccf20c49ecfe6d7618b
PiperOrigin-RevId: 170077036