Windows, aar_resources_extractor: longpath bugfix
Fix a bug in the Android tool
aar_resources_extractor: open the compressed data
and the target file as file descriptores, rather
than using aar.extract.

This way we can shorten the directories in the
relative path of the extracted file, because we
can open the target file via a junction and thus
work around long path limitations.

//tools/android:aar_resources_extractor_test now
passes on Windows.

Change-Id: I9f5c7de2279ad8769e53e015b8334a58a76cd989
PiperOrigin-RevId: 170181143