Fix flakiness of BuildEventServiceTests
The flakiness was caused by waiting for the server to receive the second
life-cycle event, but not waiting for the ack to get to the client. In that
case, the interruption would happen before the life-cycle call returned, in
which case it threw the InterruptedException, and then the stream doesn't
exist yet, and so there's no mechanism to send anything to the server.

PiperOrigin-RevId: 170472383