build_event_stream_test: fix usage of ed(1)
Unfortunately the various ed(1)s are different in subtle ways. The difference
that hit us here is the different treatment of the implicit 'e' command when ed
is started with a file as argument; GNU ed writes the answer of that command
(i.e., the number of bytes read) on stderr, whereas ed on FreeBSD writes it on
stdout. So, let's merge the two output channels and throw away the first answer
explicitly.

Fixes #3842.

Change-Id: I4ef394957cdc316f896b54ca218d4f118465caff
PiperOrigin-RevId: 170506490