Windows,launcher: add LauncherFileWriteAction
Introduce LauncherFileWriteAction, a
FileWriteAction that can create a native
{java,sh,py}_binary launcher from the launcher
stub and the user-specified launch information.

The LauncherFileWriteAction class does not use the
"copy" command of cmd.exe so it's not restricted
with path lengths.

The class stores a LaunchInfo object, which
describes the launch data that the launcher stub
reads to identify its payload.

The LaunchInfo is a lazy object, similar to
CustomCommandLine. LaunchInfo won't fully
construct the launch data until it is about to
write the data to the output. Thus LaunchInfo is
memory efficient because it won't create any
String objects in the analysis phase that it would
only read in the execution phase.

Fixes https://github.com/bazelbuild/bazel/issues/3802

Change-Id: I4ddd83369e7131d42e2e9b35f105ad2dc60bcc52
PiperOrigin-RevId: 171115105