Add .imports file to the LTOBackendAction inputs
Building with --experimental_discard_orphaned_artifacts will cause the .imports
file produced by the LTO indexing action to not be preserved, unless it is marked
as an input to the LTOBackendAction. It is only used in the input discovery phase,
not as an input to the actual action, so it was not initially listed as an input.

RELNOTES: None
PiperOrigin-RevId: 171387388