Only delete the current execroot in clean.
I noticed a problem where if you have a workspace with basename
"server", clean will rudely delete the server's pid file and cause it
to commit suicide. This is because clean deletes the deep and non-deep
execroot, presumably temporarily as part of the deep execroot
migration. --deep_execroot has been enabled for more than a year now,
so hopefully we can safely remove this aggressive deleting. --expunge
can take care of old execroots if needed.

Change-Id: I445b0d7cedf2fb9a6a365eacc85b75428a981640
PiperOrigin-RevId: 171948100