Accept command lines from tools invoking Bazel.
For tools that wrap Bazel in some way, the original way that the tool was invoked can be a useful piece of information to track when logging what Bazel did and why.

In order to output this information in the same way that Bazel outputs its command lines, we accept --tool_command_line in the structure command line format that Bazel uses in the BEP. These structured command lines are protos that we expect as a base64 encoded byte array. For simple scripts that wish to use this feature without compiling the proto, we will also accept any old string (that cannot be interpreted as a base64 encoding) as a single "chunk" in a structured command line.

This is experimental for now and users should not get attached to the format. We will remove the experimental_ prefix when it is stable.

RELNOTES: None.
PiperOrigin-RevId: 172341216