Windows: genenrate a fake CROSSTOOL when VC not found
When Visual C++ build tools are not found on Windows, we generating a
fake CROSSTOOL so that users can still build for other languages.
In the fake CROSSTOOL, we point all tools to a batch script which prints
error messages to notify users when they try to build any cc targets.

Also, added an extra default_toolchain specification for x64_windows cpu
value, so that users can use this cpu in select on non-Windows
platforms.

Fixed https://github.com/bazelbuild/bazel/issues/3853

Change-Id: Ib4d7a38fd4d98075f6b73ce310c28822e58af4e2
PiperOrigin-RevId: 172341985