Stream rest cache file uploads.
This means we no longer keep large action inputs or outputs in memory
during upload.

I adjusted the SimpleBlogStore interface to require clients to pass in
the InputStream length. That allows us to always set Content-length on
uploads. It's polite to do so, so that the server may, e.g.,
preallocate space for the blob.

Fixes https://github.com/bazelbuild/bazel/issues/3250.

Change-Id: I944c9dbc35fa2fa80dce523b0133ea9757bb3973
PiperOrigin-RevId: 172433522