Skylint: don't warn for unused initialization with 'None'.
It's good practice to place a "declaration" of a variable in a location that dominates all its uses, especially so if you want to document the variable. Hence we want no warning saying this declaration is unused.

Example:
var = None # don't warn about the unused declaration here
if condition:
  var = 0
else:
  var = 1

As noted in the code, this also has the undesired side-effect of suppressing the warning here:
var = None # doesn't warn either but ideally should
var = 0

Fixing the latter problem is not worth the increased code complexity, in my opinion.

RELNOTES: none
PiperOrigin-RevId: 172442402