Release notes: correctly strip out none, n/a, no
The previous regex was only matching lines and was not even correctly
doing so. We no match everything that is starting with none, n/a or
no as a sentence or as a full line. We also match empty string and
'.' as a full line.

RELNOTES: .
RELNOTES: None.
RELNOTES: No.

Change-Id: I6e85627170ce35e652c90a37e735b7db93c88b4e
PiperOrigin-RevId: 172452278