Set the platform name in the build event stream to the CPU instead of the toolchain identifier.
This is necessary because we don't want BuildConfiguration to depend on BUILD and CROSSTOOL files anymore and this is possible because the toolchain ID was conceived as a unique identifier for the configuration in the build event stream, which it isn't anyway, and this way it's at least more human-readable.

RELNOTES: None.
PiperOrigin-RevId: 172725386