Trim function params in TemplateExpander
There is a MacOS integration test that checks that we're trimming before
looking up labels, but it only checks for genrules, which aren't using this
code path yet.

This was changed in https://github.com/bazelbuild/bazel/commit/1a8d6b888ce532f5e053ecab37874740ce61e37f and we could see breakages in non-genrule
usage if a release picks up that change but not this.

PiperOrigin-RevId: 173393718