Implement android_binary.instruments attribute.
android_binary now takes an android_binary in the `instruments` attribute. If the attribute is explicitly specified, an AndroidInstrumentationInfo provider is included in the RuleConfiguredTarget, which android_instrumentation_test can use.

Having the provider allows android_instrumentation_test to check and fail if it tries to depend on an android_binary that is not instrumenting anything.

RELNOTES: None.
PiperOrigin-RevId: 173547240