Clean up LocalResourceContainer.Builder
LocalResourceContainer.Builder is referenced a bunch of times, but most of
those calls are identical. Replace them with a couple of factory methods. This
lets us make the builder private.

Additionally, LocalResourceContainer called attributeError() on the 'resources'
attribute, even though 'resource_files' is now used most of the time instead.
Now, the error is always given for the correct attribute.

RELNOTES: none
PiperOrigin-RevId: 174038035