Schedule LtoBackendAction when creating LtoBackendArtifact
There is no need to defer the action creation and scheduling, which
happens shortly thereafter, and just requires saving a bunch of information
on the LtoBackendArtifact.

This is preliminary restructuring that will aid some larger changes for
making tests that use static linking more efficient when built with ThinLTO
(for b/67424063).

RELNOTES: None
PiperOrigin-RevId: 174490283