Tweak java runfiles resolution.
Building a binary rule does not generally ensure that the runfiles
trees of its transitive data dependencies are up-to-date, so it's
generally a bad idea to use a dependency's runfiles tree from the
primary target. See https://github.com/bazelbuild/bazel/issues/3919
for an example of the problems this can cause.

This CL ensures that we pick the primary target's runfiles tree when
executing a java_binary from a runfiles tree.

This makes the Java runfiles path resolution similar to that of the
Python rules. (See commit [1].)

[1] 58ee85afcab07374dabc5493c780cbe3369b644f ("Don't follow symlink when looking for python module space")

Change-Id: I412ede5cf02ab2c407e45a2b262442ca67df9ba6
PiperOrigin-RevId: 174501597