Make the config expansion behavior modular.
Reorganize the BlazeOptionHandler to make it easy to exchange out how the config options are expanded using the definitions in the blazerc.

Also change the getOptionsMap call to actually structure the rc options in the order that we parse them: we expand them in command order (for the test command, first add all "common" options, then "build," then "test") and then within each command, we expand the options in the rc order. This somewhat simplifies the rc expansion code, and avoids the two-phase regrouping that used to happen.

This change should not change the semantics of rc option ordering. A followup change will add an alternative implementation for --config's expansion.

RELNOTES: None.
PiperOrigin-RevId: 175208971