Update java.bzl best practices regarding HTTPS
While HTTP is faster and SHA256 already guaranteed data
hasn't been tampered with, it does give visibility over
the wire regarding which jars are being transmitted. The
true purpose of this change is to bring these practices
into greater consistency with Google's broader vision of
using HTTPS as much as possible.

PiperOrigin-RevId: 175328286