Allow @Ignore on junit test classes
Currently a test class annotated with `@Ignore` will cause the test
runner to fail with

```
Exception in thread "main" java.lang.IllegalArgumentException: Top test must be a suite
	at com.google.testing.junit.runner.junit4.JUnit4TestModelBuilder.get(JUnit4TestModelBuilder.java:53)
```

This change treats classes with no tests (either no @Test annotations or
@Ignore at class level) as an empty test suite. The main motivation
behind this is allowing an entire test class to be ignored (e.g. to
quickly deal with a flaky test) without having to modify the BUILD file.
This is desirable in order to reduce the likelihood that a developer
forgets to update the BUILD file when removing the `@Ignore` annotation.

This pull request overrides the previous pull request
https://github.com/bazelbuild/bazel/pull/4073