bootstrap test: also verify SOURCE_DATE_EPOCH
Extend our integration test that verifies that bazel can
successfully be bootstrapped from the distribution artifact,
to also verify that SOURCE_DATE_EPOCH is honored. As the
bazel binary created contains a time stamp, we need to support
it being set from the invocation to ensure a reproducible
build. The standard way of setting time stamps to be embedded
is setting the SOURCE_DATE_EPOCH environment variable. So ensure
that that time stamp occurs in the version output.

Change-Id: Ib6388104d83426b6a74d995c324adee586560fe7
PiperOrigin-RevId: 175546459