UI: show status instead of strategy for non-executing actions
Even if an action is planned to run and a strategy assigned,
it does not mean it is actually executing; e.g., it might be
waiting for local resources. To reduce confusion, show the
status instead of the execution strategy for non-executing
actions. Also avoid the word "running" for the total number
of executing and scheduled to be executed actions.

Improves on Issue #4089.

Change-Id: If658c1a24ee26eb27ccd892847af18015355a8d3
PiperOrigin-RevId: 175818071